trace: Fix format warning
Fix the following format warning by changing the length type from size_t
to int:

lib/trace/trace_writer_local.cpp: In member function ‘void trace::LocalWriter::open()’:
lib/trace/trace_writer_local.cpp:163:44: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 2 has type ‘size_t’ {aka ‘long unsigned int’} [-Wformat=]
  163 |         szFileName = os::String::format("%.*s%s%s", length, lpFileName, suffix, ext ? ext : "");
      |                                          ~~^~       ~~~~~~
      |                                            |        |
      |                                            int      size_t {aka long unsigned int}

Fixes: 4d15c6dde5e7 ("Don't make timestamp and output options mutually exclusive")

+++ lib/trace/trace_writer_local.cpp
@@ -158,7 +158,7 @@ LocalWriter::open(void) {
         }
     } else if (hasTimestamp) {
         const char *ext = strrchr(lpFileName, '.');
-        size_t length = ext ? ext - lpFileName : strlen(lpFileName);
+        int length = ext ? ext - lpFileName : strlen(lpFileName);
 
         szFileName = os::String::format("%.*s%s%s", length, lpFileName, suffix, ext ? ext : "");
         lpFileName = szFileName;
